# backend/openapi_v3.yaml
openapi: 3.0.1
info:
  title: 团队心流引擎 (TeamFlow Engine) API
  description: 为《团队心流引擎》应用提供后端服务的RESTful API。该API支持团队管理、每日状态签到、数据可视化和AI分析等核心功能。
  version: 1.2.1
...
paths:
...
  /dashboard/teams/{team_id}/compass/:
    get:
      summary: 获取团队情绪罗盘数据
      deprecated: false
      description: ''
      operationId: getCompassData
      tags:
        - Dashboard
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
        - name: period
          in: query
          description: ''
          required: false
          schema:
            type: string
            enum:
              - day
              - week
              - month
            default: week
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompassData'
          headers: {}
      security: []
  /dashboard/teams/{team_id}/focus-time/:
    get:
      summary: 获取团队有效专注时长数据
      deprecated: false
      description: ''
      operationId: getFocusTimeData
      tags:
        - Dashboard
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
        - name: period
          in: query
          description: ''
          required: false
          schema:
            type: string
            enum:
              - day
              - week
              - month
            default: week
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FocusTimeData'
          headers: {}
      security: []
  /dashboard/teams/{team_id}/insights/:
    get:
      summary: 获取AI洞察墙数据
      deprecated: false
      description: ''
      operationId: getInsightsData
      tags:
        - Dashboard
      parameters:
        - name: team_id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
            format: uuid
        - name: period
          in: query
          description: ''
          required: false
          schema:
            type: string
            enum:
              - day
              - week
              - month
            default: week
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AIInsights'
          headers: {}
      security: []
...
components:
  schemas:
    FocusTimeData:
      type: object
      properties:
        period:
          type: string
        total_hours:
          type: number
        daily_trend:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              hours:
                type: number
    AIInsights:
      type: object
      properties:
        boosters_wordcloud:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: integer
        blockers_wordcloud:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              value:
                type: integer
...
    CompassData:
      type: object
      properties:
        period:
          type: string
        trend_data:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              avg_challenge:
                type: number
              avg_skill:
                type: number
        distribution:
          type: object
          additionalProperties:
            type: number
          properties: {}
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers: []
security: []